requirements.txt
import mysql.connector
from mysql.connector import Error
uvicorn
fastapi[standard]
typing
pydantic
mysql-connector-python
# pymongo
# dnspython
# mysqlclient

mysql-connector-python

